SearchFragment: show service name in search hint (#13)
* SearchFragment: show service name in search hint
In this commit put the name of the service in the hint.

* Show search filter in search bar hint

* SearchFragment: show filter in brackets behind service name

This is still not perfect, but it will show the selected search filter
in addition to the service name, like: “Search YouTube (Playlists)”.

It will not distinguish between a YouTube Music and Youtube filter, so
it will display the same thing. Could be improved, but then the text
gets too long! :(

Co-authored-by: Profpatsch <mail@profpatsch.de>
Co-authored-by: Stypox <stypox@pm.me>