Improve skipping silences.
ExoPlayer allows controlling silence skipping by adjusting the silence duration
that is removed by `SilenceSkippingAudioProcessor`.
The default behavior with the default parameters for this component makes it
hard to listen to content because it leaves almost no pauses. After this commit
it is possible to adjust silence length that remains and shorten only long
pauses.