[PlayQueue] Fix incorrect UI states of PlayQueue items
Author: TobiGr
onNext() is called after onSubscribe() when creating a PlayQueueAdapter. For that reason the last broadcasted event is applied to the UI state although it is already reflected in the PlayQueue that was used to initialize the adapter.

This is the intended behavior of the previously used event broadcaster of the type BehaviorSubject<>. The broadcaster's type was changed to PublishSubject<> which does not emit the last event after onSubscribe().

Co-Authored-By: Tobi <17365767+tobigr@users.noreply.github.com>